added skipTrailingZeroes() helper function
int skipTrailingZeroes counts through every trailing zero in the float C String's mantissa. It returns the location of the first number in the mantissa that is not a zero.
Having a helper function for skipping zeroes puts less work on the mantissa function by reducing the amount of time it takes for the mantissa C String to be converted into an integer numerator.